Move common/net/connection.go into cnc folder
