+++ common/net/cnc/connection.go
@@ -1,12 +1,12 @@
-package net
+package cnc
 
 import (
 	"io"
-	"net"
 	"time"
 
 	"github.com/xtls/xray-core/common"
 	"github.com/xtls/xray-core/common/buf"
+	"github.com/xtls/xray-core/common/net"
 	"github.com/xtls/xray-core/common/signal/done"
 )
 
@@ -88,8 +88,8 @@ type connection struct {
 	writer  buf.Writer
 	done    *done.Instance
 	onClose io.Closer
-	local   Addr
-	remote  Addr
+	local   net.Addr
+	remote  net.Addr
 }
 
 func (c *connection) Read(b []byte) (int, error) {
